requirements.txt
flask==3.0.3
torch==2.6.0
torchvision==0.21.0
timm==1.0.9
Pillow==10.4.0
numpy==1.26.4